requirements.txt
opencv-python-headless
numpy
pytesseract
Pillow
pytest
